Role assignment and Authorization

+++ Controllers/SessionsController.cs
@@ -18,7 +18,7 @@ namespace StudyGroups.Controllers
         // GET: Sessions
         public ActionResult Index()
         {
-            var sessions = db.Sessions.Include(s => s.Subject);
+            var sessions = db.Sessions.Include(s => s.Creator).Include(s => s.StudyGroup);
             return View(sessions.ToList());
         }
 
@@ -38,31 +38,48 @@ namespace StudyGroups.Controllers
         }
 
         // GET: Sessions/Create
+        [Authorize(Roles = "User")]
         public ActionResult Create()
         {
-            ViewBag.SubjectID = new SelectList(db.Subjects, "SubjectID", "Title");
+            //study groups created by the user only
+            int currentUserID = (int)Session["UserID"];
+            ViewBag.StudyGroupID = new SelectList(
+                db.StudyGroups.Where(sg => sg.CreatorUserID == currentUserID),
+                "StudyGroupUD",
+                "Name"
+                );
             return View();
         }
 
         // POST: Sessions/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "SessionID,Date,Duration,SubjectID")] Session session)
+        [Authorize(Roles = "User")]
+        public ActionResult Create([Bind(Include = "SessionID,Date,Duration,StudyGroupID,CreatorUserID")] Session session)
         {
             if (ModelState.IsValid)
             {
+                session.CreatorUserID = (int)Session["UserID"];
+
                 db.Sessions.Add(session);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
 
-            ViewBag.SubjectID = new SelectList(db.Subjects, "SubjectID", "Title", session.SubjectID);
+            int currentUserID = (int)Session["UserID"];
+            ViewBag.StudyGroupID = new SelectList(
+                db.StudyGroups.Where(sg => sg.CreatorUserID == currentUserID),
+                "StudyGroupID",
+                "Name",
+                session.StudyGroupID
+             );
+
             return View(session);
         }
 
+
         // GET: Sessions/Edit/5
+        [Authorize]
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -74,48 +91,107 @@ namespace StudyGroups.Controllers
             {
                 return HttpNotFound();
             }
-            ViewBag.SubjectID = new SelectList(db.Subjects, "SubjectID", "Title", session.SubjectID);
+
+            // check if the current logged in user is the creator
+            int currentUserID = (int)Session["UserID"];
+            if (session.CreatorUserID != currentUserID)
+            {
+                TempData["Error"] = "You can only edit sessions you created. ";
+                return RedirectToAction("Details", new { id = id });
+            }
+
+            ViewBag.StudyGroupID = new SelectList(
+               db.StudyGroups.Where(sg => sg.CreatorUserID == currentUserID),
+               "StudyGroupID",
+               "Name",
+               session.StudyGroupID
+           );
+
             return View(session);
         }
 
         // POST: Sessions/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "SessionID,Date,Duration,SubjectID")] Session session)
+        [Authorize]
+        public ActionResult Edit([Bind(Include = "SessionID,Date,Duration,StudyGroupID,CreatorUserID")] Session session)
         {
+            int currentUserId = (int)Session["UserID"];
+            var originalSession = db.Sessions.FirstOrDefault(s => s.SessionID == session.SessionID);
+
+            if (originalSession == null)
+            {
+                return HttpNotFound();
+            }
+
+            if (originalSession.CreatorUserID != currentUserId)
+            {
+                TempData["Error"] = "You can only edit sessions you created.";
+                return RedirectToAction("Index");
+            }
+
             if (ModelState.IsValid)
             {
                 db.Entry(session).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            ViewBag.SubjectID = new SelectList(db.Subjects, "SubjectID", "Title", session.SubjectID);
+
+            ViewBag.StudyGroupID = new SelectList(
+                db.StudyGroups.Where(sg => sg.CreatorUserID == currentUserId),
+                "StudyGroupID",
+                "Name",
+                session.StudyGroupID
+            );
+
             return View(session);
         }
 
         // GET: Sessions/Delete/5
+        [Authorize]
         public ActionResult Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             Session session = db.Sessions.Find(id);
             if (session == null)
             {
                 return HttpNotFound();
             }
+
+            int currentUserId = (int)Session["UserID"];
+            if (session.CreatorUserID != currentUserId)
+            {
+                TempData["Error"] = "You can only delete sessions you created.";
+                return RedirectToAction("Details", new { id = id });
+            }
+
             return View(session);
         }
 
         // POST: Sessions/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult DeleteConfirmed(int id)
         {
             Session session = db.Sessions.Find(id);
+
+            if (session == null)
+            {
+                return HttpNotFound();
+            }
+
+            int currentUserId = (int)Session["UserID"];
+            if (session.CreatorUserID != currentUserId)
+            {
+                TempData["Error"] = "You can only delete sessions you created.";
+                return RedirectToAction("Index");
+            }
+
             db.Sessions.Remove(session);
             db.SaveChanges();
             return RedirectToAction("Index");
