+++ Controllers/StudyGroupsController.cs
@@ -18,7 +18,8 @@ namespace StudyGroups.Controllers
         // GET: StudyGroups
         public ActionResult Index()
         {
-            return View(db.StudyGroups.ToList());
+            var studyGroups = db.StudyGroups.Include(s => s.Creator).Include(s => s.Subject);
+            return View(studyGroups.ToList());
         }
 
         // GET: StudyGroups/Details/5
@@ -37,29 +38,35 @@ namespace StudyGroups.Controllers
         }
 
         // GET: StudyGroups/Create
+        [Authorize(Roles = "User")]
         public ActionResult Create()
         {
+            ViewBag.SubjectID = new SelectList(db.Subjects, "SubjectID", "Title");
             return View();
         }
 
         // POST: StudyGroups/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "StudyGroupID,Name,Description")] StudyGroup studyGroup)
+        [Authorize(Roles = "User")]
+        public ActionResult Create([Bind(Include = "StudyGroupID,Name,Description,SubjectID,CreatorUserID")] StudyGroup studyGroup)
         {
             if (ModelState.IsValid)
             {
+                //the creator is the current logged in user
+                studyGroup.CreatorUserID = (int)Session["UserID"];
+
                 db.StudyGroups.Add(studyGroup);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
 
+            ViewBag.SubjectID = new SelectList(db.Subjects, "SubjectID", "Title", studyGroup.SubjectID);
             return View(studyGroup);
         }
 
         // GET: StudyGroups/Edit/5
+        [Authorize]
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -71,46 +78,97 @@ namespace StudyGroups.Controllers
             {
                 return HttpNotFound();
             }
+
+            // check if the current logged in user is the creator
+            int currentUserID = (int)Session["UserID"];
+            if (studyGroup.CreatorUserID != currentUserID)
+            {
+                TempData["Error"] = "You can only edit study groups you created.";
+                return RedirectToAction("Details", new { id = id });
+            }
+
+            ViewBag.SubjectID = new SelectList(db.Subjects, "SubjectID", "Title", studyGroup.SubjectID);
             return View(studyGroup);
         }
 
         // POST: StudyGroups/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "StudyGroupID,Name,Description")] StudyGroup studyGroup)
+        [Authorize]
+        public ActionResult Edit([Bind(Include = "StudyGroupID,Name,Description,SubjectID,CreatorUserID")] StudyGroup studyGroup)
         {
+            //verify ownership
+            int currentUserID = (int)Session["UserID"];
+            var originalGroup = db.StudyGroups.FirstOrDefault(sg => sg.StudyGroupID == studyGroup.StudyGroupID);
+
+            if (originalGroup == null)
+            {
+                return HttpNotFound();
+            }
+
+            if (originalGroup.CreatorUserID != currentUserID)
+            {
+                TempData["Error"] = "You can only edit study group you created.";
+                return RedirectToAction("Index");
+            }
+
             if (ModelState.IsValid)
             {
                 db.Entry(studyGroup).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
+            ViewBag.SubjectID = new SelectList(db.Subjects, "SubjectID", "Title", studyGroup.SubjectID);
             return View(studyGroup);
         }
 
         // GET: StudyGroups/Delete/5
+        [Authorize]
         public ActionResult Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             StudyGroup studyGroup = db.StudyGroups.Find(id);
             if (studyGroup == null)
             {
                 return HttpNotFound();
             }
+
+            //check if the current user is the creator
+            int currentUserID = (int)Session["UserID"];
+            if (studyGroup.CreatorUserID != currentUserID)
+            {
+                TempData["Error"] = "You can only delete study groups you created.";
+                return RedirectToAction("Details", new {id = id});
+            }
+
             return View(studyGroup);
         }
 
         // POST: StudyGroups/Delete/5
+        [Authorize]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
             StudyGroup studyGroup = db.StudyGroups.Find(id);
+
+            if (studyGroup == null) 
+            {
+                return HttpNotFound();
+            }
+
+            //check if the current user is the creator
+            int currentUserID = (int)Session["UserID"];
+            if (studyGroup.CreatorUserID != currentUserID)
+            {
+                TempData["Error"] = "You can only delete study groups you created. ";
+                return RedirectToAction("Index");
+            }
+
             db.StudyGroups.Remove(studyGroup);
             db.SaveChanges();
             return RedirectToAction("Index");
